Decouple queue
